✨ (DiscogsModels.swift): introduce new Discogs models for search response and release details to streamline data handling ♻️ (DiscogsService.swift): refactor DiscogsService to utilize new models and improve error handling for API interactions 📝 (DiscogsService.swift): remove outdated models and comments to enhance code clarity and maintainability
